Move Server-side ChatSystem Regions to Partial Classes (#44092)
* chore: move regions to partial classes

* chore: forgot to remove the region
